% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeFormats.R
\name{dateTimeFormats}
\alias{dateTimeFormats}
\title{dateTimeFormats}
\usage{
dateTimeFormats()
}
\value{
No return values. Instead a matrix is displayed with all the options for date-time objects, with examples.
}
\description{
dateTimeFormats displays the possible date-time options available (with examples).
}
\examples{
# Run this function at any time you want to.
dateTimeFormats()
}
