% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeFormats2.R
\name{dateTimeFormats2}
\alias{dateTimeFormats2}
\title{dateTimeFormats2}
\usage{
dateTimeFormats2()
}
\value{
No return values. Instead a vector is displayed with all the options for date-time objects.
}
\description{
dateTimeFormats displays the possible date-time options available (without examples).
}
\examples{
# Run this function at any time you want to.
dateTimeFormats2()
}
