% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esAssigned.R
\docType{data}
\name{esAssigned}
\alias{esAssigned}
\title{Resulting dataset after function esAssign has been executed on the raw ESM dataset.}
\format{An object of class \code{list} of length 4.}
\usage{
esAssigned
}
\description{
\strong{Note}: The list of datasets \code{esAssigned} is the result of the function \code{\link{esAssign}}, when the raw ESM dataset \code{esMerged1} is one of the function arguments. \code{esMerged1} is the result of function \code{\link{convertChars}}.
}
\keyword{datasets}
