% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esMerged1.R
\docType{data}
\name{esMerged1}
\alias{esMerged1}
\title{Resulting dataset after function convertChars has been executed on the raw ESM dataset.}
\format{An object of class \code{data.frame} with 228 rows and 32 columns.}
\usage{
esMerged1
}
\description{
\strong{Note}: The dataset \code{esMerged1} is the result of the function \code{\link{convertChars}}, when the raw ESM dataset \code{esMerged} is one of the function arguments. \code{esMerged} is the result of function \code{\link{esMerge}}.
}
\keyword{datasets}
