% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedDf.R
\docType{data}
\name{expectedDf}
\alias{expectedDf}
\title{Resulting dataset after function expectedPromptIndex has been executed on the raw ESM dataset.}
\format{An object of class \code{data.frame} with 225 rows and 54 columns.}
\usage{
expectedDf
}
\description{
\strong{Note}: The dataset \code{expectedDf} is the result of the function \code{\link{expectedPromptIndex}}, when the raw ESM dataset \code{madeShift$esDf} is one of the function arguments. \code{madeShift$esDf} is one of the results of function \code{\link{makeShift}}.
}
\keyword{datasets}
