% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCompleteLs.R
\docType{data}
\name{isCompleteLs}
\alias{isCompleteLs}
\title{Resulting dataset after function esComplete has been executed on the list of ESM datasets.}
\format{An object of class \code{list} of length 7.}
\usage{
isCompleteLs
}
\description{
\strong{Note}: The dataset \code{isCompleteLs} is the result of the function \code{\link{esComplete}}, when the list of datasets \code{rmInvLs[["dfValid"]]} is one of the function arguments. \code{rmInvLs[["dfValid"]]} is one of the results of function \code{\link{rmInvalid}}.
}
\keyword{datasets}
