% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyLsNew.R
\docType{data}
\name{keyLsNew}
\alias{keyLsNew}
\title{Resulting dataset after function genDateTime has been executed on the list of ESM datasets.}
\format{An object of class \code{list} of length 6.}
\usage{
keyLsNew
}
\description{
\strong{Note}: The dataset \code{keyLsNew} is the result of the function \code{\link{genDateTime}}, when the list of datasets \code{keyLs} is one of the function arguments.
}
\keyword{datasets}
