% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceDfNew.R
\docType{data}
\name{referenceDfNew}
\alias{referenceDfNew}
\title{Dataset 'referenceDf' in modified form}
\format{An object of class \code{data.frame} with 8 rows and 12 columns.}
\usage{
referenceDfNew
}
\description{
\strong{Note}: The dataset \code{referenceDfNew} is the result of the function \code{\link{genDateTime}}, when the dataset \code{\link{referenceDf}} is one of the function arguments.
}
\examples{
# Convert the 15 digit IMEI number from scientfic notation to text.
referenceDfNew$imei <- as.character(referenceDfNew$imei)
# Display the whole dataset in the console
referenceDfNew
}
\keyword{datasets}
