% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RELEVANTINFO_ES.R
\docType{data}
\name{RELEVANTINFO_ES}
\alias{RELEVANTINFO_ES}
\title{Resulting list after function setES has been executed.}
\format{An object of class \code{list} of length 3.}
\usage{
RELEVANTINFO_ES
}
\description{
\strong{Note}: The list \code{RELEVANTINFO_ES} is the result of the function \code{\link{setES}}.
}
\keyword{datasets}
