% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RELEVANTVN_ES.R
\docType{data}
\name{RELEVANTVN_ESext}
\alias{RELEVANTVN_ESext}
\title{Resulting list after function genDateTime has been executed on the list of ESM datasets.}
\format{An object of class \code{list} of length 8.}
\usage{
RELEVANTVN_ESext
}
\description{
\strong{Note}: The extended (see ending 'ext') list \code{RELEVANTVN_ESext} is the result of the function \code{\link{genDateTime}}, when the list of datasets \code{keyLs} is one of the function arguments. The list \code{RELEVANTVN_ES} (before being extended) is generated by the function \code{\link{setES}}.
}
\keyword{datasets}
