% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RELEVANTVN_REF.R
\docType{data}
\name{RELEVANTVN_REFext}
\alias{RELEVANTVN_REFext}
\title{Resulting list after function genDateTime has been executed on the reference dataset.}
\format{An object of class \code{list} of length 9.}
\usage{
RELEVANTVN_REFext
}
\description{
\strong{Note}: The extended (see ending 'ext') list \code{RELEVANTVN_REFext} is the result of the function \code{\link{genDateTime}}, when the reference dataset \code{\link{referenceDf}} is one of the function arguments. The list \code{RELEVANTVN_REF} (before being extended) is generated by the function \code{\link{setREF}}.
}
\keyword{datasets}
