% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbsqDf.R
\docType{data}
\name{tbsqDf}
\alias{tbsqDf}
\title{Resulting dataset after function computeTimeBetween has been executed on the raw ESM dataset.}
\format{An object of class \code{data.frame} with 186 rows and 59 columns.}
\usage{
tbsqDf
}
\description{
\strong{Note}: The dataset \code{tbsqDf} is the result of the function \code{\link{computeTimeBetween}}, when the raw ESM dataset \code{durDf} is one of the function arguments. \code{durDf} is the result of function \code{\link{computeDuration}}.
}
\keyword{datasets}
